message_ids should be u32s not u16s due to wraparound risk

+++ crates/core/src/transport/connection_handler.rs
@@ -470,7 +470,7 @@ pub(super) struct RemoteConnection {
     remote_is_gateway: bool,
     remote_addr: SocketAddr,
     inbound_packet_sender: mpsc::Sender<PacketData>,
-    last_message_id: u16,
+    last_message_id: u32,
 }
 
 // Define a custom error type for the transport layer
